+++ app/src/main/java/com/guru/composecookbook/ui/cryptoappmvvm/data/db/entities/Crypto.kt
@@ -16,7 +16,7 @@ data class Crypto(
     val high: Double,
     val low: Double,
     val marketCap: Long,
-    val volume: Long,
+    val volume: Double,
     val supply: Double?
 ) : Serializable {
